[Input] Fixed a bug where the camera position could become NaN. [D3D11] Fixed a crash where when running in debug mode while the DirectX SDK was not installed, the engine would fail to created a device.

+++ Runtime/Rendering/Renderer_Passes.cpp
@@ -965,7 +965,7 @@ namespace Spartan
         }
 
 		// Upsample + blend
-        for (int i = m_render_tex_bloom.size() - 1; i > 0; i--)
+        for (int i = static_cast<int>(m_render_tex_bloom.size() - 1); i > 0; i--)
         {
             upsample(m_render_tex_bloom[i], m_render_tex_bloom[i - 1]);
         }
